Disable SSRF redirect tests
